+++ attributes/default.rb
@@ -26,12 +26,31 @@ default['iptables-ng']['enabled_ip_versions'] = [4, 6]
 # necessary to remove the "nat" and "raw" tables.
 default['iptables-ng']['enabled_tables'] = %w(nat filter mangle raw)
 
+# Configure whether the service should be managed by Chef
+# /!\ Be careful when using this feature as it might leave your server in an inconsistent state.
+# See https://github.com/chr4-cookbooks/iptables-ng/pull/42 for more details.
+default['iptables-ng']['managed_service'] = true
+
+# Enable nat support for ipv6
+# Older distributions do not support ipv6 nat, but recent Ubuntu does
+default['iptables-ng']['ip6tables_nat_support'] = value_for_platform(
+  'ubuntu' => { '14.04' => true, '14.10' => true, 'default' => false },
+  'default' => false,
+)
+
 # Packages to install
 default['iptables-ng']['packages'] = case node['platform_family']
 when 'debian'
   %w(iptables iptables-persistent)
 when 'rhel'
-  %w(iptables iptables-ipv6)
+  if node['platform'] == 'amazon'
+    # Amazon Linux doesn't include "iptables-services" or "iptables-ipv6"
+    %w(iptables)
+  elsif node['platform_version'].to_f >= 7.0
+    %w(iptables iptables-services)
+  else
+    %w(iptables iptables-ipv6)
+  end
 else
   %w(iptables)
 end
@@ -42,20 +61,31 @@ when 'debian'
   # Debian squeeze (and before) only support an outdated version
   # of iptables-persistent, which is not capable of ipv6.
   # Furthermore, restarting the service doesn't properly reload the rules
+  if node['platform_version'].to_f >= 8.0
+    default['iptables-ng']['service_ipv4'] = 'netfilter-persistent'
+    default['iptables-ng']['service_ipv6'] = 'netfilter-persistent'
+  elsif node['platform_version'].to_f >= 7.0
+    default['iptables-ng']['service_ipv4'] = 'iptables-persistent'
+    default['iptables-ng']['service_ipv6'] = 'iptables-persistent'
+  end
+
   if node['platform_version'].to_f < 7.0
     # default['iptables-ng']['service_ipv4'] = 'iptables-persistent'
     default['iptables-ng']['script_ipv4'] = '/etc/iptables/rules'
     default['iptables-ng']['script_ipv6'] = '/etc/iptables/rules.v6'
   else
-    default['iptables-ng']['service_ipv4'] = 'iptables-persistent'
-    default['iptables-ng']['service_ipv6'] = 'iptables-persistent'
     default['iptables-ng']['script_ipv4'] = '/etc/iptables/rules.v4'
     default['iptables-ng']['script_ipv6'] = '/etc/iptables/rules.v6'
   end
 
 when 'ubuntu'
-  default['iptables-ng']['service_ipv4'] = 'iptables-persistent'
-  default['iptables-ng']['service_ipv6'] = 'iptables-persistent'
+  if node['platform_version'].to_f >= 14.10
+    default['iptables-ng']['service_ipv4'] = 'netfilter-persistent'
+    default['iptables-ng']['service_ipv6'] = 'netfilter-persistent'
+  else
+    default['iptables-ng']['service_ipv4'] = 'iptables-persistent'
+    default['iptables-ng']['service_ipv6'] = 'iptables-persistent'
+  end
   default['iptables-ng']['script_ipv4'] = '/etc/iptables/rules.v4'
   default['iptables-ng']['script_ipv6'] = '/etc/iptables/rules.v6'