+++ resources/rule.rb
@@ -21,7 +21,9 @@
 actions        :create, :create_if_missing, :delete
 default_action :create
 
-attribute :name,       kind_of: String, name_attribute: true
+# Allow only dashes, underscores word characters and dots in rule name
+# As the name attribute will be used as a filename later
+attribute :name,       kind_of: String, name_attribute: true, regex: /^[-\w\.]+$/
 # linux/netfilter/x_tables.h doesn't restrict chains very tightly.  Just a string token
 # with a max length of XT_EXTENSION_MAXLEN (29 in all 3.x headers I could find)
 attribute :chain,      kind_of: String, default: 'INPUT',  regex: /^[\w-]{1,29}$/
@@ -35,4 +37,18 @@ def initialize(*args)
 
   # Include iptables-ng::install recipe
   @run_context.include_recipe('iptables-ng::install')
+end
+
+def path_for_chain
+  "/etc/iptables.d/#{table}/#{chain}"
+end
+
+def path_for_ip_version(version)
+  "#{path_for_chain}/#{name}.rule_v#{version}"
+end
+
+def paths
+  Array(ip_version).map do |version|
+    path_for_ip_version version
+  end
 end