+++ spec/default_spec.rb
@@ -17,7 +17,7 @@ describe 'iptables-ng::default' do
         chain:      'INPUT',
         table:      'filter',
         rule:       '--protocol tcp --dport 22 --match state --state NEW --jump ACCEPT',
-        ip_version: [4, 6]
+        ip_version: [4, 6],
       )
 
     expect(chef_run).to create_iptables_ng_rule('test2-filter-INPUT-attribute-rule')
@@ -25,7 +25,7 @@ describe 'iptables-ng::default' do
         chain:      'INPUT',
         table:      'filter',
         rule:       '--protocol tcp --dport 80 --match state --state NEW --jump DROP',
-        ip_version: 4
+        ip_version: 4,
       )
   end
 end