+++ test/cookbooks/iptables_ng_test/files/default/tests/minitest/lwrp_rule_create_custom_chain_test.rb
@@ -11,8 +11,12 @@ describe 'iptables-ng::lwrp_create_custom_chain' do
     file('/etc/iptables.d/nat/FOO/custom-chain-output.rule_v4').must_include('--protocol icmp --jump ACCEPT')
   end
 
-  it 'should not set custom ip6tables rule for nat chain' do
-    file('/etc/iptables.d/nat/FOO/custom-chain-output.rule_v6').wont_exist
+  it 'should only set custom ip6tables rule for nat chain if ip6tables_nat_support attribute is enabled' do
+    if node['iptables-ng']['ip6tables_nat_support']
+      file('/etc/iptables.d/nat/FOO/custom-chain-output.rule_v6').must_include('--protocol icmp --jump ACCEPT')
+    else
+      file('/etc/iptables.d/nat/FOO/custom-chain-output.rule_v6').wont_exist
+    end
   end
 
   it 'should enable iptables serices' do